requirements.txt
av==12.3.0
djitellopy==2.5.0
numpy==2.1.0
opencv-python==4.10.0.84
pillow==10.4.0